% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.r
\name{account_history}
\alias{account_history}
\title{Get history associated with your account.}
\usage{
account_history(source = c("", "web", "api"), dt = NULL, type = c("",
  "search", "classify", "tag", "watch"), focus = NULL,
  auth = passive_auth())
}
\arguments{
\item{source}{optional    Source of the action: \code{web}, \code{api}}

\item{dt}{optional    Datetime to be used as a filter.}

\item{type}{optional Type of history: \code{search}, \code{classify}, \code{tag}, \code{watch}}

\item{focus}{optional Specific value that was used as the focus of the history.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}

\item{Account}{functions}
}
\value{
\code{tbl_df} of account history details
}
\description{
Get history associated with your account.
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Account-GetV2AccountHistory}
}

