% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.r
\name{account_info}
\alias{account_info}
\title{Get account details your account.}
\usage{
account_info(auth = passive_auth())
}
\arguments{
\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}

\item{Account}{functions}
}
\value{
\code{tbl_df} of account details
}
\description{
Get account details your account.
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Account-GetV2Account}
}

