% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.r
\name{account_notifications}
\alias{account_notifications}
\title{Get notifications that have been posted to your account.}
\usage{
account_notifications(type = NULL, auth = passive_auth())
}
\arguments{
\item{type}{optional Type of notifications to get back.}

\item{Account}{functions}
}
\value{
\code{tbl_df} of account notifications details
}
\description{
Get notifications that have been posted to your account.
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Account-GetV2AccountNotifications}
}

