% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.r
\name{account_organization}
\alias{account_organization}
\title{Get details about the organization your account is associated with.}
\usage{
account_organization(auth = passive_auth())
}
\arguments{
\item{type}{optional Type of notifications to get back.}

\item{Account}{functions}
}
\value{
\code{list} of account organization details
}
\description{
Get details about the organization your account is associated with.
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Account-GetV2AccountOrganization}
}

