% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.r
\name{account_sources}
\alias{account_sources}
\title{Get source details for a specific source.}
\usage{
account_sources(source = NULL, auth = passive_auth())
}
\arguments{
\item{source}{optional Name of the source to pull back.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{list} of account source details
}
\description{
Get source details for a specific source.
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(account_sources)} to invalidate the cache for this
      function.
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Account-GetV2AccountSourcesSource}
}
\seealso{
Other Account functions: \code{\link{account_history}},
  \code{\link{account_info}},
  \code{\link{account_notifications}},
  \code{\link{account_organization}},
  \code{\link{account_teamstream}}
}

