% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.r
\name{account_teamstream}
\alias{account_teamstream}
\title{Get the teamstream for the organization your account is associated with.}
\usage{
account_teamstream(source = c("", "web", "api"), dt = NULL, type = c("",
  "search", "classify", "tag", "watch"), focus = NULL,
  auth = passive_auth())
}
\arguments{
\item{source}{optional    Source of the action: \code{web}, \code{api}}

\item{dt}{optional    Datetime to be used as a filter.}

\item{type}{optional Type of history: \code{search}, \code{classify}, \code{tag}, \code{watch}}

\item{focus}{optional Specific value that was used as the focus of the history.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{tbl_df} of account teamstream details
}
\description{
Get the teamstream for the organization your account is associated with.
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(account_teamstream)} to invalidate the cache for this
      function.
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Account-GetV2AccountOrganizationTeamstream}
}
\seealso{
Other Account functions: \code{\link{account_history}},
  \code{\link{account_info}},
  \code{\link{account_notifications}},
  \code{\link{account_organization}},
  \code{\link{account_sources}}
}

