% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passivetotal-package.r
\name{forget}
\alias{forget}
\title{Flush the cache for a given API function}
\description{
By default, \code{passivetotal} API functions cache results using
    the \code{memoise} R pacakge. There are times when you need to ensure you
    are receiving up-to-date information. By calling \code{forget} with the bare function
    name from the \code{passivetotal} package you can invalidate the cache for
    that function/API call.
}
\examples{
\dontrun{
forget(passive_dns)
}
}

