% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve metadata about a domain or IP address}
\usage{
get_metadata(domain_or_ip)
}
\arguments{
\item{domain_or_ip}{atomic character vector containing an IP address or
domain name}
}
\value{
list of values
}
\description{
Metadata describes the item being queried and includes many of the options
available inside of the action API calls.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\examples{
\dontrun{
get_metadata("www.passivetotal.com")
$ever_compromised
[1] FALSE

$tags
list()

$dynamic
[1] FALSE

$value
[1] "www.passivetotal.com"

$subdomains
list()

$query_value
[1] "www.passivetotal.com"

$tld
[1] ".com"

$primaryDomain
[1] "passivetotal.com"

$type
[1] "domain"
}
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

