% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{get_passive}
\alias{get_passive}
\title{Get dynamic status of domain}
\usage{
get_passive(domain_or_ip)
}
\arguments{
\item{domain_or_ip}{atomic character vector containing an IP address or
domain name}
}
\value{
list
}
\description{
PassiveTotal provides a complete passive DNS picture for a domain or IP
address including first/last seen values, deconflicted values, sources used,
unique counts and enrichment for all values. In order to save bandwidth,
enrichment data is stored in it's own dictionary value and not associated
with the records.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\examples{
\dontrun{
get_passive("107.170.89.121")
$sources_used
[1] "dnsres"      "virustotal"  "mnemonic"    "domaintools" "pingly"      "kaspersky"   "alienvault"

$records
Source: local data frame [108 x 5]

      source          resolve           last_seen          first_seen                                                      record_hash
1  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46 03f3f72de51b2af8bfa940c05b4a24bad81b2e4b741a7a017af140e63fb5cdc8
2  kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30 aaf5b53667756d0e4d9ec2877ccb8ea18c71ce26a5f2038a8d747d38ed733eef
3  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46 03f3f72de51b2af8bfa940c05b4a24bad81b2e4b741a7a017af140e63fb5cdc8
4  kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30 ff43de05bc17d5e71e24f634db25598839916b5ca59913ddd2d6f086462e0cb1
5  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46 03f3f72de51b2af8bfa940c05b4a24bad81b2e4b741a7a017af140e63fb5cdc8
6  kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30 aaf5b53667756d0e4d9ec2877ccb8ea18c71ce26a5f2038a8d747d38ed733eef
7  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46 03f3f72de51b2af8bfa940c05b4a24bad81b2e4b741a7a017af140e63fb5cdc8
8  kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30 4469bd446b6c548e4f947565782b9b7152aabe5ce40818e2382792c0eec28779
9  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46 03f3f72de51b2af8bfa940c05b4a24bad81b2e4b741a7a017af140e63fb5cdc8
10 kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30 aaf5b53667756d0e4d9ec2877ccb8ea18c71ce26a5f2038a8d747d38ed733eef
..       ...              ...                 ...                 ...                                                              ...

$unique_resolutions
                domain  n
1 www.passivetotal.org 22
2 www.passivetotal.com 17
3     passivetotal.com 32
4     passivetotal.org 37

$enrichment_map
$enrichment_map$www.passivetotal.org
$enrichment_map$www.passivetotal.org$ever_compromised
[1] FALSE

$enrichment_map$www.passivetotal.org$classification
[1] ""

$enrichment_map$www.passivetotal.org$tags
list()

$enrichment_map$www.passivetotal.org$dynamic
[1] FALSE

$enrichment_map$www.passivetotal.org$value
[1] "www.passivetotal.org"

$enrichment_map$www.passivetotal.org$subdomains
list()

$enrichment_map$www.passivetotal.org$tld
[1] ".org"

$enrichment_map$www.passivetotal.org$primaryDomain
[1] "passivetotal.org"

$enrichment_map$www.passivetotal.org$type
[1] "domain"


$enrichment_map$www.passivetotal.com
$enrichment_map$www.passivetotal.com$primaryDomain
[1] "passivetotal.com"

$enrichment_map$www.passivetotal.com$classification
[1] ""

$enrichment_map$www.passivetotal.com$tags
list()

$enrichment_map$www.passivetotal.com$dynamic
[1] FALSE

$enrichment_map$www.passivetotal.com$value
[1] "www.passivetotal.com"

$enrichment_map$www.passivetotal.com$subdomains
list()

$enrichment_map$www.passivetotal.com$tld
[1] ".com"

$enrichment_map$www.passivetotal.com$ever_compromised
[1] FALSE

$enrichment_map$www.passivetotal.com$type
[1] "domain"


$enrichment_map$passivetotal.com
$enrichment_map$passivetotal.com$ever_compromised
[1] FALSE

$enrichment_map$passivetotal.com$classification
[1] ""

$enrichment_map$passivetotal.com$tags
list()

$enrichment_map$passivetotal.com$dynamic
[1] FALSE

$enrichment_map$passivetotal.com$value
[1] "passivetotal.com"

$enrichment_map$passivetotal.com$subdomains
list()

$enrichment_map$passivetotal.com$tld
[1] ".com"

$enrichment_map$passivetotal.com$primaryDomain
[1] "passivetotal.com"

$enrichment_map$passivetotal.com$type
[1] "domain"


$enrichment_map$passivetotal.org
$enrichment_map$passivetotal.org$ever_compromised
[1] FALSE

$enrichment_map$passivetotal.org$classification
[1] ""

$enrichment_map$passivetotal.org$tags
list()

$enrichment_map$passivetotal.org$dynamic
[1] FALSE

$enrichment_map$passivetotal.org$value
[1] "passivetotal.org"

$enrichment_map$passivetotal.org$subdomains
list()

$enrichment_map$passivetotal.org$tld
[1] ".org"

$enrichment_map$passivetotal.org$primaryDomain
[1] "passivetotal.org"

$enrichment_map$passivetotal.org$type
[1] "domain"



$basic_map
Source: local data frame [4 x 7]

    primary_domain classification dynamic                value  tld ever_compromised   type
1 passivetotal.org                  FALSE www.passivetotal.org .org            FALSE domain
2 passivetotal.com                  FALSE www.passivetotal.com .com            FALSE domain
3 passivetotal.com                  FALSE     passivetotal.com .com            FALSE domain
4 passivetotal.org                  FALSE     passivetotal.org .org            FALSE domain
}
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

