% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{get_subdomains}
\alias{get_subdomains}
\title{Get subdomains for a domain}
\usage{
get_subdomains(domain)
}
\arguments{
\item{domain}{atomic character vector containing a domain name}
}
\description{
Subdomains provides a comprehensive view of all known subdomains for a
registered domain with associated passive DNS information. This call is best
used to understand the activity of a particular domain over a period of
time. Passive DNS information is only deconflicted at the subdomain level,
not across the entire domain.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

