% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{get_tags}
\alias{get_tags}
\title{Get user tags for a domain or IP address}
\usage{
get_tags(domain_or_ip)
}
\arguments{
\item{domain_or_ip}{atomic character vector containing an IP address or
domain name}
}
\value{
tbl_df
}
\description{
PassiveTotal uses three types of tags (user, global, and temporal) in order
to provide context back to the user.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

