% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{is_dynamic_associated}
\alias{is_dynamic_associated}
\title{Get whether a domain is associated with a dynamic DNS provider}
\usage{
is_dynamic_associated(domain)
}
\arguments{
\item{domain}{atomic character vector containing a domain name}
}
\value{
logical
}
\description{
PassiveTotal allows users to notate if a domain is associated with a
dynamic DNS provider.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

