% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is_fqdn}
\alias{is_fqdn}
\title{Validate that a string looks like a fully qualified domain name}
\usage{
is_fqdn(x)
}
\arguments{
\item{x}{string to test}
}
\value{
\code{logical}
}
\description{
Returns \code{FALSE} if \code{x} does not "look like" a FQDN
}
\examples{
is_fqdn("rud.is")
}
\seealso{
Other Utility functions: \code{\link{is_ipv4}}
}

