% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is_ipv4}
\alias{is_ipv4}
\title{Validate that a string is an IPv4 address}
\usage{
is_ipv4(x)
}
\arguments{
\item{x}{string to test}
}
\value{
\code{logical}
}
\description{
Returns \code{FALSE} if \code{x} is not four octet IPv4 address
}
\examples{
is_ipv4("10.10.10.10")
}
\seealso{
Other Utility functions: \code{\link{is_fqdn}}
}

