% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{is_on_watchlist}
\alias{is_on_watchlist}
\title{Get whether a domain is associated with a dynamic DNS provider}
\usage{
is_on_watchlist(domain_or_ip)
}
\arguments{
\item{domain_or_ip}{atomic character vector containing an IP address or
domain name}
}
\value{
logical
}
\description{
PassiveTotal allows users to notate if a domain is associated with a
dynamic DNS provider.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

