% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{is_sinkhole}
\alias{is_sinkhole}
\title{Check sinkhole status of an IP address}
\usage{
is_sinkhole(ip)
}
\arguments{
\item{ip}{atomic character vector containing an IP address}
}
\value{
logical
}
\description{
PassiveTotal allows users to notate if an IP address is a known sinkhole.
These values are shared globally with everyone in the platform.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

