% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{passive_auth}
\alias{passive_auth}
\title{Get or set PASSIVETOTAL_USER & PASSIVETOTAL_API_KEY values}
\usage{
passive_auth(force = FALSE)
}
\arguments{
\item{force}{force setting a new PassiveTotal API username & key for the
current environment?}
}
\value{
\code{list} containing the combined PassiveTotal username & API key
}
\description{
The API wrapper functions in this package all rely on a PassiveTotal API
key residing in the environment variables code{PASSIVETOTAL_USER} &
\code{PASSIVETOTAL_API_KEY}. The easiest way to accomplish this is to set
them in the `.Renviron` file in your home directory.
}

