% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.r
\name{passive_host}
\alias{passive_host}
\title{Get detailed information about a host}
\usage{
passive_host(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}
}
\value{
\code{list}
}
\description{
Get detailed information about a host
}
\examples{
\dontrun{
passive_host("passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Host_Attributes}
}
\seealso{
Other Host attribute functions: \code{\link{passive_tracker_search}},
  \code{\link{passive_tracker}}
}

