% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssl.r
\name{passive_ssl_history}
\alias{passive_ssl_history}
\title{Get the SSL certificate history associated with an IP address or SHA-1}
\usage{
passive_ssl_history(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{tbl_df}
}
\description{
Get the SSL certificate history associated with an IP address or SHA-1
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(passive_ssl_history)} to invalidate the cache for this
      function.
}
\examples{
\dontrun{
passive_ssl_history("52.8.228.23")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-SSL_Certificates-GetV2SslCertificateHistoryQuery}
}
\seealso{
Other SSL functions: \code{\link{passive_ssl_certificate}},
  \code{\link{passive_ssl_search}}
}

