% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.r
\name{passive_tag_search}
\alias{passive_tag_search}
\title{Search for items based on tag value}
\usage{
passive_tag_search(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{list}
}
\description{
Search for items based on tag value
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(passive_tag_search)} to invalidate the cache for this
      function.
}
\examples{
\dontrun{
passive_tag_search("passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Actions-GetV2ActionsTagsQuery}
}
\seealso{
Other action functions: \code{\link{passive_classification}},
  \code{\link{passive_status}}, \code{\link{passive_tags}}
}

