% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.r
\name{passive_tags}
\alias{passive_tags}
\title{Get the tags for a query value}
\usage{
passive_tags(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}
}
\value{
\code{list}
}
\description{
Get the tags for a query value
}
\examples{
\dontrun{
passive_tags("passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Actions-GetV2ActionsTagsQuery}
}
\seealso{
Other action functions: \code{\link{passive_classification}},
  \code{\link{passive_status}},
  \code{\link{passive_tag_search}}
}

