% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.r
\name{passive_tracker_search}
\alias{passive_tracker_search}
\title{Get hosts matching a specific tracker ID}
\usage{
passive_tracker_search(query, type, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{type}{A valid tracker type to search. One of
\code{YandexMetricaCounterId}, \code{ClickyId},
\code{GoogleAnalyticsAccountNumber}, \code{NewRelicId}, \code{MixpanelId},
\code{GoogleAnalyticsTrackingId}}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{tbl_df}
}
\description{
Get hosts matching a specific tracker ID
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(passive_tracker_search)} to invalidate the cache for this
      function.
}
\examples{
\dontrun{
passive_tracker_search(query="UA-49901229", type="GoogleAnalyticsAccountNumber")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Host_Attributes-GetV2HostAttributesTrackersQuery}
}
\seealso{
Other Host attribute functions: \code{\link{passive_host}},
  \code{\link{passive_tracker}}
}

