% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.r
\name{passive_unique}
\alias{passive_unique}
\title{Get unique resolutions from passive DNS data}
\usage{
passive_unique(query, direction = c("", "next", "previous"), page = NULL,
  sources = NULL, start = NULL, end = NULL, timeout = NULL,
  auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{direction}{optional Pagination direction: \code{next} or \code{previous}}

\item{page}{optional    Page ID to request}

\item{sources}{optional Comma seperated list of sources to process with}

\item{start}{optional Only show data starting from date: \code{YYYY-mm-dd}}

\item{end}{optional Only show data up to date: \code{YYYY-mm-dd}}

\item{timeout}{optional Timeout to use on all source requests}
}
\value{
\code{list}
}
\description{
Get unique resolutions from passive DNS data
}
\examples{
\dontrun{
passive_unique("passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-DNS-GetV2DnsPassiveUniqueQuery}
}
\seealso{
Other DNS functions: \code{\link{passive_dns}}
}

