% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whois.r
\name{passive_whois}
\alias{passive_whois}
\title{Get WHOIS data for a domain or IP address}
\usage{
passive_whois(query, compact = TRUE, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{compact}{Compress the WHOIS record into a deduplicated format.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{list}
}
\description{
Get WHOIS data for a domain or IP address
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(passive_whois)} to invalidate the cache for this
      function.
}
\examples{
\dontrun{
passive_whois("passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-WHOIS-GetV2WhoisQuery}
}
\seealso{
Other WHOIS functions: \code{\link{passive_whois_search}}
}

