% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whois.r
\name{passive_whois_search}
\alias{passive_whois_search}
\title{Get WHOIS records based on field matching queries.}
\usage{
passive_whois_search(query, field, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{field}{WHOIS field to execute the search on: \code{domain},
\code{email}, \code{name}, \code{organization}, \code{address},
\code{phone}, \code{nameserver}}
}
\value{
\code{list}
}
\description{
Get WHOIS records based on field matching queries.
}
\examples{
\dontrun{
passive_whois_search(query="domains@riskiq.com", field="email")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-WHOIS-GetV2WhoisQuery}
}
\seealso{
Other WHOIS functions: \code{\link{passive_whois}}
}

