% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passivetotal-package.r
\docType{package}
\name{passivetotal}
\alias{passivetotal}
\alias{passivetotal-package}
\title{Tools for Working with the (Version 2) 'PassiveTotal' API}
\description{
Provides programmatic access to the 'PassiveTotal' API
\url{https://www.passivetotal.org/api/docs/} (Version 2) which contains rich
information on internet domain names, 'IP' addresses, 'SSL' certificates and
'malware'.
}
\details{
You should set `PASSIVETOTAL_USER` & `PASSIVETOTAL_API_KEY` in `.Renviron`
or you'll either be prompted for them or will need to pass them to each
function manually.
}
\author{
Bob Rudis (@hrbrmstr)
}
\references{
\url{https://api.passivetotal.org/api/docs/}
}

