% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{remove_tag}
\alias{remove_tag}
\title{Remove a tag fgrom a domain name or IP address}
\usage{
remove_tag(domain_or_ip, tag)
}
\arguments{
\item{domain_or_ip}{a domain or IP address to tag}

\item{tag}{value used to tag query value. Should only consist of alphanumeric,
underscores and hyphen values}
}
\description{
PassiveTotal uses three types of tags (user, global, and temporal) in order
to provide context back to the user.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

