% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{set_classification}
\alias{set_classification}
\title{Set a classification for a domain or IP address}
\usage{
set_classification(domain_or_ip, classification)
}
\arguments{
\item{domain_or_ip}{atomic character vector containing an IP address or
domain name}

\item{classification}{classification to set ("targeted", "crime",
"multiple", or "benign")}
}
\description{
PassiveTotal uses the notion of classifications to highlight table rows a
certain color based on how they have been rated. There are four types of
classification: "targeted", "crime", "multiple", or "benign"
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

