% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{set_compromised}
\alias{set_compromised}
\title{Set ever_compromised status for a domain or IP address}
\usage{
set_compromised(domain_or_ip, is_compromised)
}
\arguments{
\item{domain_or_ip}{atomic character vector containing an IP address or
domain name}

\item{is_compromised}{logical (\code{TRUE} if compromised)}
}
\description{
PassiveTotal allows users to notate if a domain or IP address have ever been
compromised. These values aid in letting users know that a site may be
benign, but it was used in an attack at some point in time.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

