% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{set_dynamic_associated}
\alias{set_dynamic_associated}
\title{Set dynamic status of domain}
\usage{
set_dynamic_associated(domain_or_ip, is_dynamic)
}
\arguments{
\item{domain_or_ip}{atomic character vector containing an IP address or
domain name}

\item{is_dynamic}{logical \code{TRUE} if dynamic}
}
\description{
PassiveTotal allows users to notate if a domain is associated with a
dynamic DNS provider.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

