% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{set_watching}
\alias{set_watching}
\title{Set whether a domain is associated with a dynamic DNS provider}
\usage{
set_watching(domain_or_ip, is_watching)
}
\arguments{
\item{domain_or_ip}{atomic character vector containing an IP address or
domain name}

\item{is_watching}{logical (\code{TRUE} if watching entity)}
}
\description{
PassiveTotal allows users to "watch" domains or IP addresses in order to get
notified of any changes.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

