% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.r
\name{account_info}
\alias{account_info}
\title{Get account details your account.}
\usage{
account_info(auth = passive_auth())
}
\arguments{
\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{tbl_df} of account details
}
\description{
Get account details your account.
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(account_info)} to invalidate the cache for this
      function.
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Account-GetV2Account}
}
\seealso{
Other Account functions: \code{\link{account_history}},
  \code{\link{account_notifications}},
  \code{\link{account_organization}},
  \code{\link{account_sources}},
  \code{\link{account_teamstream}}
}

