% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.r
\name{account_notifications}
\alias{account_notifications}
\title{Get notifications that have been posted to your account.}
\usage{
account_notifications(type = NULL, auth = passive_auth())
}
\arguments{
\item{type}{optional Type of notifications to get back.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{tbl_df} of account notifications details
}
\description{
Get notifications that have been posted to your account.
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(account_notifications)} to invalidate the cache for this
      function.
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Account-GetV2AccountNotifications}
}
\seealso{
Other Account functions: \code{\link{account_history}},
  \code{\link{account_info}},
  \code{\link{account_organization}},
  \code{\link{account_sources}},
  \code{\link{account_teamstream}}
}

