% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.r
\name{account_sources}
\alias{account_sources}
\title{Get source details for a specific source.}
\usage{
account_sources(source = NULL, auth = passive_auth())
}
\arguments{
\item{source}{optional Name of the source to pull back.}

\item{Account}{functions}
}
\value{
\code{list} of account source details
}
\description{
Get source details for a specific source.
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Account-GetV2AccountSourcesSource}
}

