% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passive.R
\name{get_passive}
\alias{get_passive}
\title{Get dynamic status of domain}
\usage{
get_passive(domain_or_ip)
}
\arguments{
\item{domain_or_ip}{atomic character vector containing an IP address or
domain name}
}
\value{
list
}
\description{
PassiveTotal provides a complete passive DNS picture for a domain or IP
address including first/last seen values, deconflicted values, sources used,
unique counts and enrichment for all values. In order to save bandwidth,
enrichment data is stored in it's own dictionary value and not associated
with the records.
}
\note{
PassiveTotal API key must be set in the environment.
      See: \code{\link{passive_api_key}} for more details.
}
\examples{
\dontrun{
get_passive("107.170.89.121")
## $sources_used
## [1] "dnsres"      "virustotal"  "mnemonic"    "domaintools" "pingly"
## [6] "kaspersky"   "alienvault"
##
## $records
## Source: local data frame [120 x 5]
##
##       source          resolve           last_seen          first_seen
## 1  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46
## 2  kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30
## 3  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46
## 4  kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30
## 5  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46
## 6  kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30
## 7  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46
## 8  kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30
## 9  kaspersky passivetotal.org 2015-06-01 04:42:47 2014-07-08 14:22:46
## 10 kaspersky passivetotal.com 2015-06-01 04:42:47 2015-02-04 18:46:30
## ..       ...              ...                 ...                 ...
## Variables not shown: record_hash (chr)
##
## $unique_resolutions
##                 domain  n
## 1 www.passivetotal.org 25
## 2 www.passivetotal.com 18
## 3     passivetotal.com 35
## 4     passivetotal.org 42
##
## $basic_map
## Source: local data frame [4 x 7]
##
##     primary_domain classification dynamic                value  tld
## 1 passivetotal.org                  FALSE www.passivetotal.org .org
## 2 passivetotal.com                  FALSE www.passivetotal.com .com
## 3 passivetotal.com                  FALSE     passivetotal.com .com
## 4 passivetotal.org                  FALSE     passivetotal.org .org
## Variables not shown: ever_compromised (lgl), type (chr)
##
## $enrichment_map
## $enrichment_map$www.passivetotal.org
## $enrichment_map$www.passivetotal.org$ever_compromised
## [1] FALSE
##
## $enrichment_map$www.passivetotal.org$classification
## [1] ""
##
## $enrichment_map$www.passivetotal.org$tags
## list()
##
## $enrichment_map$www.passivetotal.org$dynamic
## [1] FALSE
##
## $enrichment_map$www.passivetotal.org$value
## [1] "www.passivetotal.org"
##
## $enrichment_map$www.passivetotal.org$subdomains
## list()
##
## $enrichment_map$www.passivetotal.org$tld
## [1] ".org"
##
## $enrichment_map$www.passivetotal.org$primaryDomain
## [1] "passivetotal.org"
##
## $enrichment_map$www.passivetotal.org$type
## [1] "domain"
##
##
## $enrichment_map$www.passivetotal.com
## $enrichment_map$www.passivetotal.com$primaryDomain
## [1] "passivetotal.com"
##
## $enrichment_map$www.passivetotal.com$classification
## [1] ""
##
## $enrichment_map$www.passivetotal.com$tags
## list()
##
## $enrichment_map$www.passivetotal.com$dynamic
## [1] FALSE
##
## $enrichment_map$www.passivetotal.com$value
## [1] "www.passivetotal.com"
##
## $enrichment_map$www.passivetotal.com$subdomains
## list()
##
## $enrichment_map$www.passivetotal.com$tld
## [1] ".com"
##
## $enrichment_map$www.passivetotal.com$ever_compromised
## [1] FALSE
##
## $enrichment_map$www.passivetotal.com$type
## [1] "domain"
##
##
## $enrichment_map$passivetotal.com
## $enrichment_map$passivetotal.com$ever_compromised
## [1] FALSE
##
## $enrichment_map$passivetotal.com$classification
## [1] ""
##
## $enrichment_map$passivetotal.com$tags
## list()
##
## $enrichment_map$passivetotal.com$dynamic
## [1] FALSE
##
## $enrichment_map$passivetotal.com$value
## [1] "passivetotal.com"
##
## $enrichment_map$passivetotal.com$subdomains
## list()
##
## $enrichment_map$passivetotal.com$tld
## [1] ".com"
##
## $enrichment_map$passivetotal.com$primaryDomain
## [1] "passivetotal.com"
##
## $enrichment_map$passivetotal.com$type
## [1] "domain"
##
##
## $enrichment_map$passivetotal.org
## $enrichment_map$passivetotal.org$ever_compromised
## [1] FALSE
##
## $enrichment_map$passivetotal.org$classification
## [1] ""
##
## $enrichment_map$passivetotal.org$tags
## list()
##
## $enrichment_map$passivetotal.org$dynamic
## [1] FALSE
##
## $enrichment_map$passivetotal.org$value
## [1] "passivetotal.org"
##
## $enrichment_map$passivetotal.org$subdomains
## list()
##
## $enrichment_map$passivetotal.org$tld
## [1] ".org"
##
## $enrichment_map$passivetotal.org$primaryDomain
## [1] "passivetotal.org"
##
## $enrichment_map$passivetotal.org$type
## [1] "domain"
}
}
\seealso{
\href{PassiveTotal API documentation}{https://www.passivetotal.org/api/docs}
}

