% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.r
\name{passive_classification}
\alias{passive_classification}
\title{Get the status of a classification domain}
\usage{
passive_classification(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{list}
}
\description{
Get the status of a classification domain
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(passive_classification)} to invalidate the cache for this
      function.
}
\examples{
\dontrun{
passive_classification("passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Actions}
}
\seealso{
Other action functions: \code{\link{passive_status}},
  \code{\link{passive_tag_search}},
  \code{\link{passive_tags}}
}

