% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{passive_enrich}
\alias{passive_enrich}
\title{Enrich the given query with metadata}
\usage{
passive_enrich(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{list}
}
\description{
Enrich the given query with metadata
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(passive_enrich)} to invalidate the cache for this
      function.
}
\examples{
\dontrun{
passive_enrich("passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Enrichment-GetV2EnrichmentQuery}
}
\seealso{
Other Enrichment functions: \code{\link{passive_malware}},
  \code{\link{passive_osint}},
  \code{\link{passive_subdomains}}
}

