% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.r
\name{passive_host}
\alias{passive_host}
\title{Get detailed information about a host}
\usage{
passive_host(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{tbl_df}
}
\description{
Get detailed information about a host
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(passive_host)} to invalidate the cache for this
      function.
}
\examples{
\dontrun{
passive_host("passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Host_Attributes}
}
\seealso{
Other Host attribute functions: \code{\link{passive_tracker_search}},
  \code{\link{passive_tracker}}
}

