% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{passive_malware}
\alias{passive_malware}
\title{Get malware data}
\usage{
passive_malware(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}
}
\value{
\code{list}
}
\description{
Get malware data
}
\examples{
\dontrun{
passive_malware("noorno.com")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Enrichment-GetV2EnrichmentMalwareQuery}
}
\seealso{
Other Enrichment functions: \code{\link{passive_enrich}},
  \code{\link{passive_osint}},
  \code{\link{passive_subdomains}}
}

