% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{passive_osint}
\alias{passive_osint}
\title{Get opensource intelligence data}
\usage{
passive_osint(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}
}
\value{
\code{list}
}
\description{
Get opensource intelligence data
}
\examples{
\dontrun{
passive_osint("xxxmobiletubez.com")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Enrichment-GetV2EnrichmentOsintQuery}
}
\seealso{
Other Enrichment functions: \code{\link{passive_enrich}},
  \code{\link{passive_malware}},
  \code{\link{passive_subdomains}}
}

