% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssl.r
\name{passive_ssl_certificate}
\alias{passive_ssl_certificate}
\title{Get the SSL certificate associated with the SHA-1.}
\usage{
passive_ssl_certificate(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{list}
}
\description{
Get the SSL certificate associated with the SHA-1.
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(passive_ssl_certificate)} to invalidate the cache for this
      function.
}
\examples{
\dontrun{
passive_ssl_certificate("e9a6647d6aba52dc47b3838c920c9ee59bad7034")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-SSL_Certificates-GetV2SslCertificateQuery}
}
\seealso{
Other SSL functions: \code{\link{passive_ssl_history}},
  \code{\link{passive_ssl_search}}
}

