% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{passive_subdomains}
\alias{passive_subdomains}
\title{Get subdomains using a wildcard query}
\usage{
passive_subdomains(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}

\item{auth}{\code{list} containing PassiveTotal \code{user} & \code{key}}
}
\value{
\code{list}
}
\description{
Get subdomains using a wildcard query
}
\note{
All API function in the \code{passivetotal} pacake use \code{memoise::memoise} to
      cache results. Call \code{forget(passive_subdomains)} to invalidate the cache for this
      function.
}
\examples{
\dontrun{
passive_subdomains("*.passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Enrichment-GetV2EnrichmentSubdomains}
}
\seealso{
Other Enrichment functions: \code{\link{passive_enrich}},
  \code{\link{passive_malware}},
  \code{\link{passive_osint}}
}

