% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.r
\name{passive_tracker}
\alias{passive_tracker}
\title{Get all tracking codes for a domain or IP address.}
\usage{
passive_tracker(query, auth = passive_auth())
}
\arguments{
\item{query}{Query value to use in your request.}
}
\value{
\code{list}
}
\description{
Get all tracking codes for a domain or IP address.
}
\examples{
\dontrun{
passive_tracker("passivetotal.org")
}
}
\references{
\url{https://api.passivetotal.org/api/docs/#api-Host_Attributes-GetV2HostAttributesTrackersQuery}
}
\seealso{
Other Host attribute functions: \code{\link{passive_host}},
  \code{\link{passive_tracker_search}}
}

