% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passivetotal-package.R
\docType{package}
\name{passivetotal}
\alias{passivetotal}
\alias{passivetotal-package}
\title{A package to interface with the PassiveTotal API}
\description{
A package to interface with the PassiveTotal API
}
\author{
Bob Rudis (@hrbrmstr)
}

